Adds chroot execution in the installation script
This commit makes an improvement to the Void Linux installation script. Chroot execution has been added to facilitate system configuration within the chroot environment. Now, after completing the main system installation, the chroot to the root directory (/mnt) is performed and an additional script is run to configure the system and users. This enhancement simplifies the installation process and provides a more automated way to perform the required configuration. In addition, minor fixes and enhancements have been made to existing scripts.